requirements.txt
aniso8601==6.0.0
asgiref==3.2.2
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
bleach==3.1.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.1
decorator==4.4.2
defusedxml==0.6.0
Django==2.1
django-filter==2.2.0
djangorestframework==3.11.1
entrypoints==0.3
Flask==1.0.2
Flask-JWT-Extended==3.18.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
idna==2.10
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
model-mommy==2.0.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
pycodestyle==2.6.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.5.2
PyMySQL==0.9.3
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.0
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.0.0
